Cleared warnings

+++ Repository/GenericRepository.cs
@@ -18,6 +18,7 @@ namespace Hatebook.Repository
         public async Task Delete(Guid id)
         {
             var entity = await _db.FindAsync(id);
+            if (entity==null) { new BadRequestObjectResult(""); return;  }
             _db.Remove(entity);
         }
 
@@ -26,7 +27,7 @@ namespace Hatebook.Repository
             _db.RemoveRange(entities);
         }
 
-        public async Task<T> Get(Expression<Func<T, bool>> expression, List<string> includes = null)
+        public async Task<T?> Get(Expression<Func<T, bool>> expression, List<string>? includes = null)
         {
             IQueryable<T> query = _db;
             if (includes != null)
@@ -39,10 +40,13 @@ namespace Hatebook.Repository
 
             return await query.AsNoTracking().FirstOrDefaultAsync(expression);
         }
-
-        public async Task<IList<T>> GetAll(Expression<Func<T, bool>> expression = null,
-            Func<IQueryable<T>, IOrderedQueryable<T>> orderBy = null,
-            List<string> includes = null)
+        Task<T> IGenericRepository<T>.Get(Expression<Func<T, bool>> expression, List<string>? includes)
+        {
+            return Get(expression, includes) as Task<T>;
+        }
+        public async Task<IList<T>> GetAll(Expression<Func<T, bool>>? expression = null,
+            Func<IQueryable<T>, IOrderedQueryable<T>>? orderBy = null,
+            List<string>? includes = null)
         {
             IQueryable<T> query = _db;
